♻️ refactor tests (which are hardly existing)
