+++ src/mqt/qudits/quantum_circuit/__init__.py
@@ -2,8 +2,9 @@
 
 from __future__ import annotations
 
-from .circuit import QuantumCircuit
+from .circuit import QuantumCircuit, QuantumRegister
 
 __all__ = [
     "QuantumCircuit",
+    "QuantumRegister",
 ]
