+++ test/python/qudits_circuits/test_z.py
@@ -1,7 +1,10 @@
+from __future__ import annotations
+
 from unittest import TestCase
 
 import numpy as np
-from mqt.qudits.qudit_circuits.circuit import QuantumCircuit
+
+from mqt.qudits.quantum_circuit import QuantumCircuit
 
 
 def omega_d(d):
@@ -19,7 +22,7 @@ class TestZ(TestCase):
 
         z_1 = self.circuit_23.z(1)
         matrix_1 = z_1.to_matrix(identities=0)
-        assert np.allclose(np.array([[1, 0, 0], [0, omega_d(3), 0], [0, 0, (omega_d(3)**2)]]), matrix_1)
+        assert np.allclose(np.array([[1, 0, 0], [0, omega_d(3), 0], [0, 0, (omega_d(3) ** 2)]]), matrix_1)
 
     def test_validate_parameter(self):
         z = self.circuit_23.z(0)
